[Profile] Creating an endpoint for fetching profile data
